Add egui template and refactor cli into a utils directory
