+++ monsterbook/src/lib.rs
@@ -1,5 +1,7 @@
 extern crate image;
 extern crate nshare;
 
+pub mod app;
 pub mod crop;
 pub mod stitch;
+pub mod utils;
